src/services/catalog/pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]

[project]
name = "catalog_service"
version = "0.1.0"
readme = "README.md"
authors = [{ name = "yeegie", email = "lotus9200@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "click==8.3.0",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.117.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "pydantic==2.11.9",
    "pydantic_core==2.33.2",
    "PyYAML==6.0.2",
    "python-slugify",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.43",
    "starlette==0.48.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.15.0",
    "uvicorn==0.36.0",
    "asyncpg"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }

[project.scripts]
catalog_service = "bin.api.main:main"

[tool.mypy]
ignore_missing_imports = true
